tools/requirements.txt
################################################################################
## AUTOGENERATED: This file is autogenerated by req-compile.
##
## Python: 3.11.14
## Platform: Darwin
##
## To regenerate this file, use the following command:
##
##    bazel run "@@//tools:requirements.update"
##
################################################################################

PyYAML==6.0.3 \
    --hash=sha256:652cb6edd41e718550aad172851962662ff2681490a8a711af6a4d288dd96824
    # via _main/tools/requirements.in (>=6.0.1)
    # https://files.pythonhosted.org/packages/16/19/13de8e4377ed53079ee996e1ab0a9c33ec2faf808a4647b7b4c0d46dd239/pyyaml-6.0.3-cp311-cp311-macosx_11_0_arm64.whl#sha256=652cb6edd41e718550aad172851962662ff2681490a8a711af6a4d288dd96824
